SCCP optimization support for all architectures, with reuse of operators/condition/addressing mode static evaluation.
Reuse as much as possible the value analysis for operators from
Compcert. ValueDomainSSA.v is just a duplicate from ValueDomain.v
to handle SSA genvs. And file $(ARCH)/ValueAOpSSA.v is a
duplicate + simplification of $(ARCH)/ValueAOp.v, to prune the
analysis of the stack pointer -- it maps to Vtop, so that the
correctness theorems are not requiring the block classification of
sp to be BCstack, but just different from BCinvalid.